+++ api/graphql/index.ts
@@ -1,5 +1,0 @@
-import { makeExecutableSchema } from 'graphql-tools';
-
-const schema = '1'
-
-export { schema };