+++ api/modules.d.ts
@@ -1,4 +1,0 @@
-declare module "*.gql" {
-    const value: any;
-    export default value;
-}