+++ config/config.ts
@@ -1,4 +1,4 @@
-import path from "path";
+import path from 'path'
 
 /**
  * Absolute path to root directory.
@@ -8,7 +8,7 @@ import path from "path";
  * So we need to make workaround.
  * 2x ../ because we are in config->gulp
  */
-export const rootPath = path.resolve(__dirname, "../");
+export const rootPath = path.resolve(__dirname, '../')
 
 /**
  * Create path to files from root path
@@ -17,18 +17,38 @@ export const rootPath = path.resolve(__dirname, "../");
  * @returns {string} resolved path from root to files
  */
 export const rootPathFunc = (files: string): string => {
-  return path.resolve(rootPath, files);
-};
+	return path.resolve(rootPath, files)
+}
+
+/**
+ * Paths for files, dir etc.
+ */
+export const paths = {
+	appDir: 'app',
+	wwwDir: 'www',
+	app: {
+		public: 'app/public',
+		views: 'app/views/**/*.html',
+		scss: 'app/public/scss',
+		ts: 'app/public/ts'
+	},
+	www: {
+		css: 'www/css',
+		img: 'www/img',
+		js: 'www/js',
+		api: 'www/api'
+	}
+}
 
 /**
  * Name of main typescript file for webpack
  */
-export const entryFile = "main.ts";
+// export const entryFile = "main.ts";
 
 /**
  * Get NODE_ENV variable from cross-env
  */
-export const NODE_ENV = process.env.NODE_ENV;
+export const NODE_ENV = process.env.NODE_ENV
 
 /**
  * Define current environment
@@ -36,20 +56,20 @@ export const NODE_ENV = process.env.NODE_ENV;
  * Depends on NODE_ENV variable which is created by cross-env
  */
 export const environment = {
-  prod: NODE_ENV === "production",
-  dev: NODE_ENV === "development"
-};
+	prod: NODE_ENV === 'production',
+	dev: NODE_ENV === 'development'
+}
 
 /**
  * Define webpack mode for config file
  */
-export const webpackMode = (environment.dev ? ('development' as 'development') : 'production' as 'production')
+export const webpackMode = environment.dev ? 'development' as 'development' : 'production' as 'production'
 
 /**
  * Define options for sass compiler
  */
 export const sassOptions = {
-  // nested, expanded, compact, compressed
-  outputStyle: environment.prod ? "compressed" : "expanded",
-  precision: 5
-};
+	// nested, expanded, compact, compressed
+	outputStyle: environment.prod ? 'compressed' : 'expanded',
+	precision: 5
+}