+++ webpack.config.ts
@@ -1,8 +1,16 @@
-import * as path from 'path';
+import { webpackMode } from './config/config';
 const nodeExternals = require('webpack-node-externals')
-import { entryFile, environment, rootPathFunc, webpackMode } from './config/config';
 
 const webpack = {
+	mode: webpackMode,
+	resolve: {
+		extensions: ['.ts', '.tsx', '.gql', '.js']
+	},
+}
+
+export const API = {
+	...webpack,
+	target: "node" as "node",
 	module: {
 		rules: [
 			{
@@ -12,27 +20,27 @@ const webpack = {
 			},
 			{
 				test: /\.(gql)$/,
-				// use: 'graphql-tag/loader',
 				use: 'raw-loader',
 				exclude: /node_modules/
 			}
 		],
 	},
-	resolve: {
-		extensions: ['.ts', '.tsx', '.gql', '.js']
-	},
 	externals: [nodeExternals()]
-}
-
-export const API = {
-	mode: webpackMode,
-	target: "node" as "node",
-	...webpack,
-	// watch: true
 };
 
 export const Client = {
-	mode: webpackMode,
-	// target: "node" as "node",
 	...webpack,
+	target: "web" as "web",
+	module: {
+		rules: [
+			{
+				test: /\.ts$/,
+				use: 'ts-loader',
+			},
+			{
+				test: /\.(gql)$/,
+				use: 'raw-loader',
+			}
+		],
+	},
 };